path-over lemmas (#451)
Adds to path over lemmas. This pr contains:

- `nat-htpy-apd-htpy`, which proves that `nat-htpy H p` corresponds to
`apd H p` under `compute-path-over-eq-value`
- `preserves-tr-apd-function`, which proves that `preserves-tr f p`
corresponds to `apd f p` under `compute-path-over-eq-value`

One thing is that the direction of commutative squares used in both
`compute-path-over` functions are the opposite direction of `nat-htpy`
and `preserves-tr`, respectively. (oops, thats my fault for not thinking
ahead). I suggest we change the definitions of the `compute-path-over`
equivalences to accomodate this. I would be happy to do that in this pr,
but I wanted to get your feedback.

Co-authored-by: Egbert Rijke <e.m.rijke@gmail.com>