Add nix support files (#472)
[Nix](https://nixos.org/) is a build tool/environment manager built with
reproducibility and isolation in mind; one of its use cases is for
declaratively describing a collection of programs, which should be
available for working on a project (a so-called `devShell`).
For `agda-unimath`, this devShell consists of the right version of Agda
and some utilities for building the website.

A new nix-enabled developer can then clone the repository, and without
fiddling with any Haskell or Agda installation run `nix develop`, which
drops them into a shell with all the tools available, and `make check`
should just work.

The option of not depending on a system-wide implicit installation of
Agda also enables developers to work on multiple projects with different
version requirements for tooling.

A good complement for nix is [direnv](https://direnv.net/) - a command
line tool which removes the need for manually calling `nix develop`,
because when you `cd` into the project's directory (or open the project
in an IDE, if it has support for direnv, e.g. see
[envrc](https://github.com/purcell/envrc) for Emacs), it automatically
loads the environment into the current shell session, restoring the
previous state when `cd`ing out.

Further down the line, the flake.nix could provide a recipe for
`agda-unimath` (a "derivation" in Nix parlance), so that projects
depending on `agda-unimath` using Nix can setup their environment with
`pkgs.agda.withPackages [ agda-unimath ]`, but that's out of scope for
this PR. For more details on the Agda ecosystem in Nix, see the [Agda
section](https://nixos.org/manual/nixpkgs/stable/#agda) in the nixpkgs
manual.