Add "suggest an edit" button and minor fix (#484)
Due to a Github bug, the Agda block inside the details tag is not
rendered correctly on the Github website if the block contains blank
lines. Fredrik filed an issue on GitHub, and it is currently being
investigated. In the meantime, we delete the blank lines in the Agda
code import section. The imports have been sorted as well.

Also, I added a "Suggest an edit" button on the website to make it
easier to suggest edits to the Agda code. The button is located at the
top right corner of the menu bar. I also fixed the Agda-UniMath title,
which now links to the correct URL.

Check this version here: https://jonaprieto.github.io/agda-unimath/