Generate module indexes (#501)
Automatically generate both namespace index modules and the main index
file (meaning the long list in `SUMMARY.md`) as part of `pre-commit`.

The namespace index modules are generated by
- Create a file if it does not exist
- If there is no title, use the namespace but capitalize the first word
and replace dashes with spaces. E.g. `orthogonal-factorization-systems`
becomes "Orthogonal factorization systems"
- If there is no Agda-block, create one, otherwise, capture the last
Agda block in the file. (this is to preserve possible `OPTION` pragmas
as in the case of `type-theories`.
- Declare the namespace module, and import all files in the namespace
folder publically and in alphabetic ordering.

This assures that if there is a custom title or other text in the
namespace index module from before (as in the case of
`univalent-combinatorics`), then this is preserved.

The main index file is generated as follows
- Overwrite the previous file
- Add title "Formalisation in Agda"
- Add entry lists for each namespace. The title of a module is used for
the entry, and the lists are sorted by title as well.

Resolves #448. 

Written in collaboration with @jonaprieto.

---------

Co-authored-by: Jonathan Cubides <jonaprieto@users.noreply.github.com>