Restore CONTRIBUTORS.md while finding a CI fix (#509)
The CI fails due to a call to the GitHub API to generate the
contributor's list. While
we find a solution for it, let's restore the corresponding file.

Btw, two additional scripts were added to an ongoing discussion
regarding the space convention so that we can keep them for the record
and experiment locally. See the history of this description and the
comments below to see all the details of why there are not running by
default.