Make `unused_imports_remover` faster and safer (#512)
- Builds on #502.

So I wanted to try using GPT-4 to improve my programs and chose the
script `unused_imports_remover` as a test case.

This PR improves the speed and safety of the program. Now, the original
Agda files are only overwritten by working code, so I deem it no longer
unsafe for others to use. Moreover, it uses multithreading and a couple
of other tricks to make the overall run time significantly shorter. To
compare, the previous version took 6.4 hours to finish on my computer
while this one only took 55 minutes.

Also gives the same treatment to `foundation_demoter`. It is now safer
and takes only 70 seconds to run.

- Also resolves #523.