Universal algebra first commit (#522)
First commit on universal algebra. Credits to @plt-amy for coming up on
how to do the `is-algebra` thing.

I'm not convinced on the names for the stuff under the sections
`Abstract equations` and `Theories`, but otherwise I think it's fine.
I think that doing
```
group-Algebra-Group (((A , is-set-A) , models-A) , satisfies-A) =
```
is not common in this library but it really helps to reduce a lot of the
boilerplate. Not sure if creating types equivalent to the standard
finite types is fine. It certainly helps readability in my opinion.