First definitions towards n-ary functoriality of set quotients (#528)
I defined n-ary operations using the functional vector approach, i.e. by
indexing with finite types. The problem with this is that then
`preserves-sim-n-ary-map-Eq-Rel` has as codomain `UUω` since agda takes
a pi type
```agda
( as : (i : Fin n) → As i) → _
```
to live in `UUω`. It seems it doesn't realize that `Fin n` is finite so
that `As i` is always bounded. This makes it so that I can't generate
the corresponding subtype of n-ary-operations.

I'd like to hear your opinions/suggestions.

---------

Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>