Final universal algebra (#544)
My final commit on universal algebra, I did not reach my goal of the
first homomorphism theorem as I've realized a lot of theory will be
needed for it, and I want to work on more immediate results.

----------------

For posterity, here's my thoughts on what and how to proof next (and
other stuff) if one wanted to continue this line
1. Prove that `induced-function-term` is the same as `eval-term` in the
sense more precise of commit
https://github.com/UniMath/agda-unimath/commit/a070945a4d5b21208f70a6ea967aac2f6dc2a066
(lazily called `assignment-vec'`).
2. Prove that the quotient algebra is indeed an algebra: use the
previous result to say that each axiom in the quotients comes from
function application in the quotient, and then prove that this function
is one that comes from functoriality in the base algebra. (Alternatively
redefine the model in the quotient algebra so that this holds by
definition).
3. (1) is hard as a lot of workarounds (mutual recursion) have to be
made to satisfy the termination checker, alternatively one could
redefine being an algebra as one that matches the
`induced-function-term`s associated to two terms `t1` and `t2`. Notice
that the de bruijn variables in each term may be different, so that an
appropriate notion of merging and extending the induced functions must
be made.
4. Once these results hold, I think classic results should be relatively
easy, including the homomorphism theorems. My goal was proving the first
homomorphism theorem and applying it to the algebraic theory of groups,
and then transporting the result to groups (through univalance). I
wanted a general and *practical* way to study and prove stuff about all
algebraic structures. In the end, I'm still not sure if this approach
would have worked...