Rational addition and format for future work on rationals (#550)
Refactored some stuff, importantly separating integer fractions from the
rationals. My idea is that all future operations on rationals be defined
on integer fractions first, and then carried to rationals; as working
with the gcd directly is horrible. I did the first part of the additions
to show how it'd look, I'd use this schema for the other operations and
properties too.

I'd like confirmation that this approach is fine, though it duplicates
the number of files.