Baire Space (#549)
I've attempted to prove Baire space is uncountable.

To do this, I need to define another notion of countability called
"directly countable". A type `X` is directly countable if there is a
surjective map `f : ℕ → X`. I proved that if `X` is inhabitied, then `X`
is countable iff `X` is directly countable.

First, I proved Baire space is a  set.
Second, using direct countability and Lawvere's fixed point theorem, I
proved Baire space is uncountable.

Note that, as a trivial fact, I proved ` ℕ` is countable. As for
equivalent definitions of countabiilty, I'm still working on it.