Cleaning up some stuff in species (#575)
In this PR I am cleaning up some of the files in and related to the
theory of species.

- I have chosen some more natural file names
- English grammar corrected
- Highlighted defined terms in `## Idea` sections
- Fixed indentation levels: Sometimes 4 space increments are used,
instead of the standard 2 space increments.
- Sometimes an entire definition was wrapped in parentheses for no
reason. I removed those parentheses.
- Refactored some constructions of equivalences: Often they are
constructed in one go, even if the construction naturally factors
through the construction of the underlying map, its inverse, and
homotopies proving that the inverse is a section and a retraction. If a
construction of an equivalence naturally factors through these steps, it
is better to turn these steps into separate definitions which can be
used later.
- Sometimes the fact that subuniverses are closed under equivalences was
reproven during a construction. There is a definition
`is-in-subuniverse-equiv` for this, so it is better and faster to use
it.
- The hypotheses on subuniverses were complicated expressions and were
repeated often. They warranted to be definitions, so I made those
definitions and factored out the complicated expressions.
- Anonymous modules should not cross over sections in a file.