Reflection and solvers (#571)
# Solvers!

Add the necessary infrastructure to use solvers and add one for
precategories (see the examples at the end of
`src/category-theory/precategory-solver.lagda.md`). There are some
important design decisions I'd like to highlight (and recieve comments
on):

- Primitive types are imported into `foundation`. As discussed with
Fredrik, maybe they could be in their own `primitive` folder. Though I'd
suggest universe levels to be there too if that would be the case.
- Since the reflection mechanism is somewhat esoteric, I've named all
the primitives just as in `Agda.Builtin`, so that it can be easierly
maintained (i.e. by just comparing with `Agda.Builtin`). This however
breaks the rule of unique identifiers, e.g. multiple constructors are
called `lit` in `src/reflection/terms.lagda.md`. Since all the modules
in `reflection` should only be used for solvers, this shouldn't cause
any problems, though.
- I've been using private blocks to add list syntax as otherwise it just
becomes unmanageable. This won't pollute the namespace though, as
they're private.

## Additional comments

- I haven't commented on what some primitives do because either it's
obvious from the name, they are very niche and probably won't be used,
or I don't know what they do.
- There is commented code
[here](https://github.com/UniMath/agda-unimath/pull/571/files#diff-cc5e1b4d5733084679a96b922f92f351a4fffb3b348c440701b19786a62e45a4R161),
which will be usable in Agda 2.6.4., but not on 2.6.3. So that I (and
others) don't forget to update it, I've left them there.
- ~~The precategories solver is not as general as it could be, it won't
work on types that compute to equations using composition and
identities. What I'd suggest in those cases is using equational
reasoning to state the equality using only composition and identities
and only then apply the solver. (Interestingly, this is a consequence of
using Sigma types instead of records, although it can be worked
around).~~ It now works with modulo beta terms! (E.g. should work with
categories too)

---------

Co-authored-by: fernando <fernando.churiv@gmail.com>
Co-authored-by: Fernando Chu <17756312+shiranaiyo@users.noreply.github.com>
Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>