Define binary operator notations for `mul-ℕ`, `mul-ℤ`, and `exp-ℕ`. (#618)
They are `_*ℕ_`, `_*ℤ_`, and `_^ℕ_` respectively.

In addition to what the title says, I've also refactored the files
- `multiplication-natural-numbers`
- `multiplication-integers`
- `exponentiation-natural-numbers`
- `absolute-value-integers`

to use the binary operators where it's natural, to get a sense for how
this would look. It should be possible to change most usages easily
enough* with a nice enough regex, but I don't want to do this before I
get some feedback on whether it is preferable.

Also, pairs like `is-injective-mul-succ-ℕ` and
`is-injective-mul-succ-ℕ'` should perhaps be renamed to something like
`is-injective-left-mul-succ-ℕ` and `is-injective-right-mul-succ-ℕ`.