Pre-commit stuff (#627)
Tried out a bunch of ideas while I was on the plane, this is mostly a
result of that.

### Summary
- Add a script `markdown_conventions.py` that enforces some markdown
conventions. This script
- Prohibits completely empty sections (I.e. when a header is immediately
succeeded by a header of the same or less deep level). Often this is a
mistake, but if a section is empty because it is left for future work, I
suggest noting this with a sentence like "This remains to be shown."
instead of leaving the section empty.
  - Prohibits top-level headers after the first line
- Fix current violators of these conventions.
- Try out an `indentation_conventions.py` script that currently only
detects unevenly indented code. I.e. when code is indented with an odd
number of spaces. There are about 1000 violations at the moment so the
script does not run by default, but it is something we can consider
enabling in the future.
- Rename some pre-commit scripts
- Define infix binary operator `_-ℤ_` for the difference of integers,
and use it where possible.

---------

Co-authored-by: VictorBlanchi <76261859+VictorBlanchi@users.noreply.github.com>
Co-authored-by: Egbert Rijke <e.m.rijke@gmail.com>
Co-authored-by: Jonathan Cubides <jonathan.cubides@uib.no>