Cycle prime decomposition is closed under cartesian product (#624)
In this pull request, I wanted to prove that the cartesian product of
the cycles prime decomposition of `n` and `m` is equal to the cycle
prime decomposition of `mul-\bN n m`. Initially, I thought it would be
easy by using the fundamental theorem of arithmetic but I encountered
some unexpected problems that I had to solve ^^.
The main idea was to 

1. Show that if we take the sorted concatenation of the lists
corresponding to the prime decomposition of `n` and `m` we obtain a
prime decomposition of `mul-\bN n m`. This was harder than I thought. I
had to show that `mul-list-\bN` is invariant by permuting the list. I
made the general case, and to do it I had to show that every
transposition (on Fin n) is a composition of adjacent transpositions.
2. Then I used the fundamental theorem of arithmetic to demonstrate that
the sorted concatenation of the lists corresponding to the prime
decomposition of `n` and `m` is equal to the prime decomposition of
`mul-\bN n m`.
3. Finally, I expected to only need to permute the cycles in the
cartesian product of the cycles prime decomposition associated to `n`
and `m` with the permutation given by the sort by insertion. I thought
that applying step 2. would be sufficient to conclude. But I had to show
this property : `map-list f (permute-list p t)` is equal to
`permute-list (map-list p f) t` and which was quite hard to prove.

---------

Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>