Towards Hasse-Weil species (#631)
I wanted to define Hasse-Weil species. To do it, I had to define some
algebraic structures on finite sets. I decided to create a **new
folder** named `finite-algebra`, because I believe it will grow sooner.
(It will be great to have the classification of finite fields for
latter). I have defined:
- Finite semigroups, monoids, commutative monoids, groups, abelian
groups, rings, commutative-finite-rings, fields and semisimple
commutative finite rings
 - Homorphisms of (commutative) finite rings
 - Products of (commutative) finite rings
 - Dependent products of (commutative) finite rings 
I have shown that for any finite set their is a finite number of ways to
put an algebraic structure on it. (except semisimple commutative finite
rings, which is harder)

---------

Co-authored-by: Egbert Rijke <e.m.rijke@gmail.com>