Make Python scripts 3.8-compatible (#634)
Python 3.8 is still the default on Ubuntu 20.04 LTS, so it seems
reasonable to support that version.

When pre-commit is first run, it detects the currently used Python
version, and uses that for running the scripts (modulo a
project-specific venv). This mechanism can be overridden by specifying a
language version in the pre-commit configuration, but that has two
drawbacks:
- Only one version has to be specified, version ranges or "minimal
supported versions" are not supported, and it's not likely that they
will be, since that's a technical limitation of the underlying venv
architecture. See the following issue for more details:
https://github.com/pre-commit/pre-commit/issues/1128 . Hence the best we
could do is specify the minimal supported version (3.9 before this PR)
as _the_ version to be used to run the hooks.
- The users needs to manage the Python installation. Pre-commit doesn't
make any effort to install the required Python version, and instead
fails with a pretty cryptic message when first trying to run the hooks,
if the user doesn't have the specific version installed. We would then
require every contributor to install Python 3.9 specifically, in
addition to what their distribution already considers the "default"
Python.

I believe that would have worse developer experience than occasionally
having a pre-commit script breaking on older systems (Python 3.8 was
released in October 2019).

Fixes #633.