Localizations and other things (#655)
### Summary
- Define localizations
- Organize sigma-closed subuniverses and reflective subuniverses
- Detect ill-formed imports blocks
- Add `unused-imports` `make` target
- Add `.editorconfig` and misc. settings tweaks. The `.editorconfig`
file allows for a few shared settings between vscode, emacs, and other
editors, making the experience of using the library slightly more
homogenous. To use the `.editorconfig` file in emacs, please install
this extension: https://github.com/editorconfig/editorconfig-emacs. It
would be beneficial to note this in the install guide as well. Maybe
someone with more emacs knowledge can do that?

Resolves #652.