Refactor nilradical proof for prime ideals (#699)
I refactored the proof that the nilradical is included in every prime
ideal as discussed in the discord server. I also added some auxiliary
functions for prime ideals viewed as radical ideals. Regarding the last
function, I wasn't sure whether to keep it because agda accepts
membership in a prime ideal and in the corresponding radical ideal as
definitionally equal, but in the nilradical proof it feels wrong to
prove something for prime ideals using a function about radical ideals
directly.
I also wasn't sure whether the naming of these functions follows
conventions correctly.

---------

Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>