Construct ring congruence from two sided ideal (#708)
The ring congruence obtained from an ideal is a congruence on the
underlying abelian group, obtained from the underlying subgroup of the
ideal. The multiplicative laws for the ideal make sure that this is also
a congruence on the underlying monoid, which is proved in the
mul-congruence-lemma. It came out a little long but that was the
simplest way I could think to do things.
This whole section could be moved to the ideals-rings module, to be more
in line with the normal-subgroups and subgroups-of-abelian-groups
modules, where their congruences are developed. I'm pretty sure there is
also an analogous equivalence between ideals and ring congruences, which
could be proved.

---------

Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>