Informal writeup of the proof that group homomorphisms are epi iff surjective (#726)
This PR only adds an informal writeup of a proof that a group
homomorphism is surjective iff it is an epi. A proof without excluded
middle can be given using the theory of acyclic maps, which I'm still
working on with Ulrik and Tom. I thought that the writeup could still be
worthwhile to have on our website.

---------

Co-authored-by: Fredrik Bakke <fredrbak@gmail.com>