Make contribution credit opt-in, maintain contributor identities
After some discussion with Egbert, we came to the conclusion that we
should make it possible for contributors to the library to choose how
they are presented on the website, without that presentation being
immediately linked to their GitHub account(s), and without needing to
rewrite the commit history in situations where the identity associated
with past commits is no longer the one a contributor wants to be
identified with.

The solution we came up with is to have a collection of contributor
identities, not a priori linked to any other service, where
individuals can set their preferred display name and "claim" git
usernames. When information obtained from the git log is
presented (currently the "Created by" page header and "Recent changes"
footer of pages), this table is consulted at website build-time to
display the preferred name.

If a contributor chooses not to be presented on the website, they can
be left out of the table.

The next step is to use this table to generate the CONTRIBUTORS.md
file from git log as well, instead of relying on GitHub APIs.