Coforks, coequalizers of types, their flattening lemma (#792)
This PR introduces coforks of parallel maps, the dependent and
non-dependent universal properties of coequalizers, the construction of
coequalizers from pushouts, and the flattening lemma for coequalizers,
asserting that sigmas commute with coequalizers.

This development mirrors the development of cocones and pushouts.

I also changed the statement of the flattening lemma for pushouts to one
that sounds more natural to me - we can construct a cocone of sigma
types from any cocone, not just a pushout; the previous definition
required the vertex to be a pushout in those definitions. Now the
statement is "if a cocone is a pushout, then the cocone derived from it
is a pushout too".