The Legendre symbol (#796)
A few things to point out:

1) The function here `legendre` does not actually care whether or not
`p` is prime, it just generally characterizes whether or not a number is
a square in `Zp` for any natural `p`. I'm not sure whether or not this
is confusing. Most interesting/important facts about the legendre symbol
leverage the fact that `p` is prime, but I felt that fact should be
invoked in those theorems, not in the definition of the function itself,
seeing as there is no type of prime numbers.

2) Most of this file is definitions and theorems about squareness in
`N`, `Z`, and `Zp`. It might be better if these are in separate files,
but I wasn't sure where they should go. These didn't exist before as far
as I could tell, and the basic justification for their existence is that
the legendre symbol needs to decide whether or not a number is a square
in` Zp`, which for `p=0` (see above) requires deciding squareness in
`Z`, which in turn requires deciding squareness in `N`. At least, that
was the most straightforward way I could think of for defining and
proving these things. Maybe at least the functions `square-Z` and
`square-Z-Mod` should be in the files where `mul-Z` and `mul-Z-Mod` are,
since `square-N` was already defined and is where `mul-N` is.

3) There's a pretty long proof of an elementary algebraic identity:
`(n+2)^2 = n(n+2) + (n+2) + (n+2)`. I didn't feel this was a general
enough fact to warrant being in the top level, so it's in a `where`
clause. There also might be a way to prove the theorem it's used for
without it, not sure.

---------

Co-authored-by: mrh <mrh57@posteo.net>