Refactor precomposition (#937)
This PR refactors precomposition, following difficulties that people
have been facing around where to put stuff involving precomposition, and
cyclic dependencies that are too easily arising. In fact, it was quite a
knot to untie. At the core of the knot was what is now called the
(dependent) universal property of equivalences, which is in its own
separate file now. I created tables, and did some minor code formatting,
but please note that this is not the main purpose of this PR.