Partial and copartial functions (#975)
This PR adds partial and copartial functions to the library. Partial
functions were needed to define partial sequences, which are ultimately
needed for some OEIS entries. Copartial functions are simply the dual of
partial functions, which I came up with while formalizing partial
functions.

Note that partial functions are defined using the open modalities, so
naturally copartial functions are defined using the closed modalities.
They specify where a function is prohibited. The terminology in the
files about copartial functions is not established in the literature (as
far as I know), and should be considered experimental.