Improve generality of homotopy whiskering operations (#976)
This PR improves the generality of the homotopy whiskering operations by
introducing more possible type dependencies into the definitions.
Moreover, it bumps the precedence of both whiskering operators, so that
in particular left whiskering takes precedence over right whiskering.
These two operators commute definitionally, so it doesn't matter which
way we associate them, and Agda is usually indecisive with which way to
print them when you ask it to. The downside to this change is that Agda
will start omitting parentheses when written together with homotopy
concatenations, so maybe this is a change for the worse.