Swap order of arguments for lifts of families of elements (#981)
Makes the arguments to definitions about lifts of elements start with
type families, and continue with the families of elements that we're
lifting. This makes sense conceptually, because e.g.
`lift-family-of-elements B` is now a function which takes a family of
elements to its type of lifts, so we can talk about concepts such as
"transport in the family of lifts" by writing `tr
(lift-family-of-elements B)`.