Proposed definition of extension types (#999)
This pull request proposes a definition for extension types, following
the discussion in #885. I think this is a useful definition to have, and
it could help easing off the load of definitions in `surjective-maps` by
being able to separate it into their own concepts, such as
`surjective-extensions-types` and friends.

This PR is intentionally kept small and is ready for review.